Use soname to create symlinks for `nvidia-340-glx-driver`
Summary: Use `objdump` to find the `soname` for libraries and create symlinks according to that. See D4521 for more details.

Test Plan: I don't have the hardware to test this driver. However, the modification in this patch was tested in other versions of nvidia driver (e.g. D4521). People with hardware compatible with this driver can go ahead and test this patch.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D4761